Guard publishers with a mutex. Related: GH-11
It is better to not let Relay have multiple goroutines publishing simultaneously
because it creates a strange interaction with publisher confirms (which publish
are you actually getting the acknowledgement for?). Instead of allowing this,
we revert to the previous behavior, and just take a lock for the duration of the
call to Publish().